Database built
Added graceful crash for the NLP functionality incase of crash during the database building, we will still recover all the data that was already in the object.
Made the range for the specified end entry partition inclusive instead of exclusive like it was before.
Added input data: FinalDoublePipeDelimited
and the output data base in json format: CoreNLPData