Readme added
Added a readme with steps in bash to follow or if one would want to edit the steps to execute the maven project that would be helpful.
Edited main for testdata to run 20 entries for quick demo purposes.
readme.sh has NON-WORKING steps to execute the maven project.
readme.sh also has steps to run in intellij and has been tested so it should work. Add problem to issues if it doesn't.